some overrides and fixes
Added a fixed version of the blazing falls map (under map 0101)

worldobjectcatalog now allows you to put the TSO catalog.xml in the objects folder much like catalog_downloads.xml
useful if you want to change it, or just want a backup.

added an override that allows you to put in clothes patch using the TSO format (gamedata/skins/example.jpg). works on avatars but not EODs currently.

The lot picker now has the ability to filter out comminuty lots and always displays them on the map using a variable loaded in worldobjectcatalog
the format used is just searching for the location on the map itself, you can add comments on the file as you wish.
you can lot location from the db using "SELECT * FROM fso_lots WHERE admit_mode="6";" and copying the locations into the clots file in the misc tuning folder.
you can then combine this with a tuning downloader (if you have one) to change it on the fly.
(I will commit this later, it needs a full re-write)

backport a fix that allows you to actually put in an apostrophe into the lot name purchase dialogue.

change to .gitignore to allow 3d objects to be commited. (commented out .obj files and commited the ones needed to build)